use auto client for scc auto
